moved aggregate_report to its own report type
